move testcases to hint director
